+++ index.html
@@ -11,8 +11,6 @@
 
 <body>
   <main>
-    <script src="js/functions.js"></script>
-    <script src="js/main.js"></script>
     <!-- Фильтрация изображений от других пользователей -->
     <section class="img-filters  img-filters--inactive  container">
       <h2 class="img-filters__title  visually-hidden">Фильтр фотографий</h2>
@@ -233,5 +231,6 @@
     <div class="img-upload__message  img-upload__message--loading">Загружаем...</div>
   </template>
 
+  <script type="module" src="./js/main.js"></script>
 </body>
 </html>
